Update to leiningen 2 and EMF 2.8.0

+++ do-release.zsh
@@ -1,12 +1,25 @@
 #!/bin/zsh
 
+echo "Unzipping jars in emf-jars/..."
+cd emf-jars/
+for x in *.jar; do
+    unzip -o $x;
+done
+cd ..
+rm -rf emf-jars/META-INF/
+
 echo "Creating an Uberjar"
 lein uberjar
 
 echo "Creating a pom.xml"
 lein pom
 
-file=`ls emf-xsd-sdk-*-standalone.jar`
+file=`ls target/emf-xsd-sdk-*-standalone.jar`
+
+if [[ ! -f ${file} ]]; then
+    echo "Error: ${file} doesn't exist!"
+    exit 1
+fi
 
 echo "Moving ${file} to ${file/-standalone/}."
 mv ${file} ${file/-standalone/}
